backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
motor==3.3.2
pymongo==4.6.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6

# Security and rate limiting
slowapi==0.1.9
redis==5.0.1

# Logging and monitoring
structlog==23.2.0

# Caching and compression
aioredis==2.0.1

# Environment variables
python-dotenv==1.0.0

# Development and testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
faker==20.1.0
factory-boy==3.3.0
pytest-xdist==3.5.0
pytest-benchmark==4.0.0
pytest-security==0.1.0